refactor code and structure
