+++ src/features/user-stats/components/stats-chart/stats-chart.styles.tsx
@@ -1,0 +1,6 @@
+import styled from "styled-components"
+
+export const Wrapper = styled.div`
+  height: 66%;
+  width: 100%;
+`
