+++ src/features/user-stats/components/stats-numbers/stats-numbers.styles.ts
@@ -1,0 +1,7 @@
+import styled from "styled-components"
+
+export const Wrapper = styled.div``
+
+export const StyledSpan = styled.p`
+  color: var(--c-extra-gray);
+`
