+++ src/features/user-stats/user-stats.tsx
@@ -1,23 +1,11 @@
-import { User } from "@firebase/auth"
-import { PokoyChartData, StatsChart } from "components/stats-chart.component"
-import { Timestamp } from "firebase/firestore"
 import { useCallback, useEffect, useState } from "react"
+import { User } from "@firebase/auth"
 import { UserSerie } from "react-charts"
-import {
-  getAverage,
-  fetchAndsetChartData,
-  getStats,
-  getTotalInHours,
-} from "./getData"
-import { StyledSpan, ChartWrapper, StatsWrapper } from "./user-stats.styles"
-
-export interface UserStatsData {
-  // TODO: remove nullable values from this field
-  firstMeditationDate: Timestamp | null
-  totalDuration: number
-  count: number
-  userId: string
-}
+import { PokoyChartData, UserStatsData } from "shared/types"
+import { fetchAndsetChartData, getStats } from "./get-data"
+import { StatsChart } from "./components/stats-chart/stats-chart.component"
+import { StatsNumbers } from "./components/stats-numbers/stats-numbers.component"
+import { Wrapper } from "./user-stats.styles"
 
 interface Props {
   user: User
@@ -38,27 +26,19 @@ export const UserStats: React.FC<Props> = ({ user }) => {
 
   const dataLength = chartData?.[0]?.data?.length
   const statsExist = dataLength && dataLength > 1
-  const totalDurationExists = !!statsData?.totalDuration
 
-  return statsExist ? (
-    <StatsWrapper>
-      {/* // TODO: extract to component */}
-      {totalDurationExists && (
+  return (
+    <Wrapper>
+      {statsExist ? (
         <>
-          <StyledSpan>
-            Total: {getTotalInHours(statsData.totalDuration)} hours
-          </StyledSpan>
-          <StyledSpan>Average: {getAverage(statsData)} minutes</StyledSpan>
+          <StatsNumbers statsData={statsData} />
+          <StatsChart pokoyData={chartData} />
         </>
+      ) : (
+        <span>
+          There are no statistics yet. Try meditating more than twice.
+        </span>
       )}
-
-      <ChartWrapper>
-        <StatsChart pokoyData={chartData} />
-      </ChartWrapper>
-    </StatsWrapper>
-  ) : (
-    <StyledSpan>
-      There are no statistics yet. Try meditating more than twice.
-    </StyledSpan>
+    </Wrapper>
   )
 }
