+++ src/shared/types/chart.types.ts
@@ -1,0 +1,6 @@
+export type PokoyChartData = {
+  // NOTE: timestamp
+  primary: Date
+  // NOTE: meditation duration
+  secondary: number
+}
